Moved contract creation forms into CreateContract directory
