+++ web-app/src/components/CreateContract/CreateContractForms/dates/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PropTypes as PT } from 'prop-types';
 import container from './container';
-import { Form, Input } from '../../common';
+import { Form, Input } from '../../../common';
 import { FormContainer } from '../styled';
 
 export const DateForm = props => {
